Remove debug, add tests for follow_graph
