+++ tests/test_anf_ir_utils.py
@@ -1,0 +1,20 @@
+from myia.anf_ir import Constant, Apply, Graph
+from myia.anf_ir_utils import dfs
+
+
+def test_dfs():
+    in0 = Constant(0)
+    in1 = Constant(1)
+    value = Apply([in0, in1], Graph())
+    assert next(dfs(value)) == value
+    assert set(dfs(value)) == {value, in0, in1}
+
+
+def test_dfs_graphs():
+    g0 = Graph()
+    in0 = Constant(g0)
+    in1 = Constant(1)
+    g0.return_ = in1
+    value = Apply([in0], Graph())
+    assert set(dfs(value)) == {value, in0}
+    assert set(dfs(value, follow_graph=True)) == {value, in0, in1}
